+++ src/documentation/constants/constants.ts
@@ -394,6 +394,8 @@ export const NAVIGATION_FOUNDATIONS = [
 ];
 
 export const CHI_VISIBLE_VERSION = [
+  '6.19.0',
+  '6.18.0',
   '6.17.0',
   '6.16.0',
   '6.15.0',